feat: highlight equal addresses
