+++ src/components/pages/home/AddressLink.tsx
@@ -1,23 +1,29 @@
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import Link from "next/link";
 import { AddressContext } from "@/components/lib/context/AddressContext";
+import { SquareArrowOutUpRight } from "lucide-react";
 
 interface AddressLinkProps {
-  type: string;
-  label: string;
+  addressClassName: string;
   address: string | null;
   href: string;
   children: React.ReactNode;
 }
 
 const AddressLink = ({
-  type,
-  label,
+  addressClassName,
   address,
   href,
   children,
 }: AddressLinkProps) => {
   const { state, dispatch } = useContext(AddressContext);
+  const [isSameOrigin, setIsSameOrigin] = useState<boolean | null>(null);
+
+  useEffect(() => {
+    const currentOrigin = window.location.origin;
+    const linkOrigin = new URL(href, currentOrigin).origin;
+    setIsSameOrigin(currentOrigin === linkOrigin);
+  }, [href]);
 
   const handleMouseEnter = () => {
     dispatch({ type: "SET_HOVERED_ADDRESS", payload: address });
@@ -27,24 +33,24 @@ const AddressLink = ({
     dispatch({ type: "SET_HOVERED_ADDRESS", payload: null });
   };
 
-  const className = `truncate ${
-    type === "l1-l2"
-      ? "max-w-28 md:max-w-48 xl:max-w-28"
-      : type === "latest-transactions" && label === "tx"
-        ? "max-w-60 md:max-w-96 xl:max-w-60"
-        : "max-w-36 md:max-w-48 xl:max-w-36"
-  }`;
-
   return (
     <Link
       href={href}
-      className={`flex items-center text-primary hover:brightness-150 ${
-        state.hoveredAddress === address ? "text-yellow-200" : ""
-      }`}
+      className={`flex items-center text-primary hover:brightness-150`}
       onMouseEnter={handleMouseEnter}
       onMouseLeave={handleMouseLeave}
     >
-      <span className={className}>{children}</span>
+      <span className={`relative p-1 ${addressClassName}`}>
+        {children}
+        <span
+          className={`pointer-events-none absolute inset-0 rounded border border-dashed border-[#FFC107] bg-[#FFC107] bg-opacity-5 transition-opacity ${
+            state.hoveredAddress === address ? "opacity-100" : "opacity-0"
+          }`}
+        ></span>
+      </span>
+      {isSameOrigin === false && (
+        <SquareArrowOutUpRight className="ml-1 size-4" />
+      )}
     </Link>
   );
 };
